Complete the direct GPU runtime and remove the eager GPU API
Finish the v20 GPU migration on the explicit CUDA Graph runtime:

- Return borrowed plan-owned outputs (GpuExecutionResult) with an
  explicit copy_output, and report device status without poisoning.
- Schedule sibling root, nested, and device-body parallel loops, and
  sequential loops inside wave templates and device bodies. Waves rebind
  every template alias of a replaced owner, Zip lanes use private
  placeholder allocations, device-body families carry their lane
  producers, and nested device loops restart their index on the device.
- Support nested conditional bodies in the native graph builder with a
  frame stack; sibling conditionals inside one body are serialized.
- Choose W and C from a geometric trial grid.
- Remove the eager GPU matrix/poly/sampler API, persistent native
  format state, the unused VRAM budget, and dead CUDA entry points.
  Rewrite the GPU benches on GpuRuntime.
- Replace the stale docs with one top-down docs/architecture.md.

+++ crates/backends/cuda/src/matrix/MatrixHash.cu
@@ -546,7 +546,7 @@ extern "C" int gpu_raw_hash_sample_emit(
         !destination || !destination->limbs || !destination->rows ||
         !destination->columns || !destination->degree ||
         destination->physical_device != plan->device ||
-        destination->degree != ctx->N ||
+        destination->degree != static_cast<uint32_t>(ctx->N) ||
         destination->limb_count != plan->moduli.size() ||
         destination_binding_base > UINT32_MAX - destination->limb_count ||
         destination->rows > UINT64_MAX / destination->columns ||
